requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
graphviz==0.20.1
python-dateutil==2.8.2
urllib3==2.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.0
pydantic==2.5.0
pydantic-settings==2.1.0